+++ src/dipdup/config.py
@@ -1473,6 +1473,7 @@ class AdvancedConfig:
     :param skip_version_check: Do not check for new DipDup versions on startup
     :param rollback_depth: A number of levels to keep for rollback
     :param crash_reporting: Enable crash reporting
+    :param decimal_precision: Adjust decimal context precision.
     """
 
     reindex: dict[ReindexingReason, ReindexingAction] = field(default_factory=dict)
@@ -1484,6 +1485,7 @@ class AdvancedConfig:
     skip_version_check: bool = False
     rollback_depth: int = 2
     crash_reporting: bool = False
+    decimal_precision: int | None = None
 
 
 @dataclass